@@ -1,10 +1,10 @@
 module github.com/gelmium/proxy-benchmark
 
-go 1.21.0
+go 1.22.3
 
 require (
-	github.com/gelmium/graceful-shutdown v0.0.2
-	github.com/gofiber/fiber/v2 v2.49.1
+	github.com/gelmium/graceful-shutdown v0.0.11
+	github.com/gofiber/fiber/v2 v2.52.4
 )
 
 require (
