@@ -1,20 +1,19 @@
-// FILEPATH: /workspaces/graphql-engine-plus/src/proxy/container_start.go
 package main
 
 import (
 	"context"
-	"math/rand"
+	"log/slog"
 	"net"
 	"os"
-	"regexp"
-	"strconv"
 	"time"
 
 	gfshutdown "github.com/gelmium/graceful-shutdown"
 	"github.com/gofiber/contrib/otelfiber"
 	"github.com/gofiber/fiber/v2"
 	"github.com/gofiber/fiber/v2/log"
+	"github.com/gofiber/fiber/v2/middleware/cors"
 	"github.com/gofiber/fiber/v2/middleware/logger"
+	"github.com/gofiber/fiber/v2/middleware/proxy"
 	jsoniter "github.com/json-iterator/go"
 	"github.com/valyala/fasthttp"
 	"go.opentelemetry.io/otel"
@@ -23,8 +22,22 @@ import (
 	oteltrace "go.opentelemetry.io/otel/trace"
 )
 
-// App Runner will always timeout after 120 seconds
+// Set a hard timeout after 110 seconds
 const UPSTREAM_TIME_OUT = 110 * time.Second
+
+// Set a hard timeout for read
+const FIBER_READ_TIME_OUT = 30 * time.Second
+
+// Set a hard timeout for write
+const FIBER_WRITE_TIME_OUT = 30 * time.Second
+
+// keep-alive is enabled by default, connection is keep alive after serving a request
+// if there is no incoming request after 120 seconds it will be closed
+const FIBER_IDLE_TIME_OUT = 120 * time.Second
+
+// The total time allowed for all process to gracefully shutdown
+// Should set to a number equal or greater than READ_TIME_OUT
+const GRACEFUL_SHUTDOWN_TIMEOUT = 30 * time.Second
 
 func waitForStartupToBeCompleted(startupCtx context.Context) {
 	if startupCtx.Err() == nil {
@@ -34,50 +47,8 @@ func waitForStartupToBeCompleted(startupCtx context.Context) {
 	}
 }
 
-var notForwardHeaderRegex = regexp.MustCompile(`(?i)^(Host|Accept-Encoding|Content-Length|Content-Type)$`)
-
 var tracer oteltrace.Tracer
 var propagators propagation.TextMapPropagator
-var otelExporter = os.Getenv("ENGINE_PLUS_ENABLE_OPEN_TELEMETRY")
-
-func setRequestHeaderUpstream(c *fiber.Ctx, agent *fiber.Agent) {
-	for k, v := range c.GetReqHeaders() {
-		// filter out the header that we don't want to forward
-		// such as: Accept-Encoding, Content-Length, Content-Type, Host
-		if notForwardHeaderRegex.MatchString(k) {
-			continue
-		}
-		agent.Set(k, v)
-	}
-	// add client IP Address to the end of the X-Forwarded-For header
-	agent.Add("X-Forwarded-For", c.IP())
-}
-
-func sendRequestToUpstream(c *fiber.Ctx, agent *fiber.Agent) error {
-	// loop through the header and set the header from the original request
-	setRequestHeaderUpstream(c, agent)
-	agent.Body(c.Body())
-	// set the timeout of proxy request
-	agent.Timeout(UPSTREAM_TIME_OUT)
-	// send the request to the upstream url using Fiber Go
-	if err := agent.Parse(); err != nil {
-		log.Error(err)
-		return c.Status(500).SendString("Internal Server Error")
-	}
-	code, body, errs := agent.Bytes()
-	if len(errs) > 0 {
-		// check if error is timeout
-		if errs[0].Error() == "timeout" {
-			return c.Status(504).SendString("Upstream GraphQL Engine Timeout")
-		}
-		// log unhandled error and return 500
-		log.Error(errs)
-		return c.Status(500).SendString("Internal Server Error")
-	}
-	// return the response from the upstream url
-	c.Set("Content-Type", "application/json")
-	return c.Status(code).Send(body)
-}
 
 var primaryEngineServerProxyClient = &fasthttp.HostClient{
 	Addr: "localhost:8881",
@@ -118,7 +89,9 @@ func processProxyResponse(c *fiber.Ctx, resp *fasthttp.Response, redisCacheClien
 	resp.Header.Del("Connection")
 	// strip other unneeded headers
 	if ttl != 0 && cacheKey > 0 && resp.StatusCode() == 200 && redisCacheClient != nil {
-		ReadResponseBodyAndSaveToCache(c.Context(), resp, redisCacheClient, ttl, familyCacheKey, cacheKey, TraceOptions{tracer, c.UserContext()})
+		// start tracer span by wrapping the context with timeout same as fiber Context timeout
+		traceCtx, _ := WrapContextCancelByAnotherContext(c.UserContext(), c.Context(), 0)
+		ReadResponseBodyAndSaveToCache(traceCtx, resp, redisCacheClient, ttl, familyCacheKey, cacheKey)
 	}
 }
 
@@ -133,9 +106,9 @@ var JsoniterConfigFastest = jsoniter.Config{
 func setupFiber(startupCtx context.Context, startupReadonlyCtx context.Context, redisCacheClient *RedisCacheClient) *fiber.App {
 	app := fiber.New(
 		fiber.Config{
-			ReadTimeout:  60 * time.Second,
-			WriteTimeout: 60 * time.Second,
-			IdleTimeout:  60 * time.Second,
+			ReadTimeout:  FIBER_READ_TIME_OUT,
+			WriteTimeout: FIBER_WRITE_TIME_OUT,
+			IdleTimeout:  FIBER_IDLE_TIME_OUT,
 			JSONEncoder:  JsoniterConfigFastest.Marshal,
 			JSONDecoder:  JsoniterConfigFastest.Unmarshal,
 			ServerHeader: "graphql-engine-plus/v1.0.0",
@@ -143,30 +116,32 @@ func setupFiber(startupCtx context.Context, startupReadonlyCtx context.Context, 
 		},
 	)
 
-	// get the PATH from environment variable
-	var v1Path = os.Getenv("ENGINE_PLUS_GRAPHQL_V1_PATH")
-	// default to /public/graphql/v1 if the env is not set
-	if v1Path == "" {
-		v1Path = "/public/graphql/v1"
-	}
-	// get the PATH from environment variable
-	var roPath = os.Getenv("ENGINE_PLUS_GRAPHQL_V1_READONLY_PATH")
-	// default to /public/graphql/v1readonly if the env is not set
-	if roPath == "" {
-		roPath = "/public/graphql/v1readonly"
-	}
-	// get the PATH from environment variable
-	var execPath = os.Getenv("ENGINE_PLUS_PUBLIC_EXECUTE_PATH")
-
 	// set logger middleware
-	if os.Getenv("DEBUG") == "true" {
+	if debugMode {
 		log.SetLevel(log.LevelDebug)
+		handler := slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelDebug})
+		slog.SetDefault(slog.New(handler))
 	} else {
 		log.SetLevel(log.LevelInfo)
+		handler := slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelInfo})
+		slog.SetDefault(slog.New(handler))
 	}
 	app.Use(logger.New(logger.Config{
 		Format:     "${time} \"${method} ${path}\" ${status} ${latency} (${bytesSent}) [${reqHeader:X-Request-ID};${reqHeader:Traceparent}] \"${reqHeader:Referer}\" \"${reqHeader:User-Agent}\"\n",
 		TimeFormat: "2006-01-02T15:04:05.000000",
+		// skip health check endpoints
+		Next: func(c *fiber.Ctx) bool {
+			// Next defines a function to skip this middleware when returned true.
+			if c.Path() == healthCheckPath || c.Path() == "/health" {
+				return true
+			}
+			return false
+		},
+	}))
+
+	// CORS middleware
+	app.Use(cors.New(cors.Config{
+		AllowOrigins: hasuraGqlCorsDomain,
 	}))
 
 	// Compress middleware, we dont need to use this
@@ -178,10 +153,10 @@ func setupFiber(startupCtx context.Context, startupReadonlyCtx context.Context, 
 	// }))
 
 	// setup opentelemetry middleware for these endpoints
-	if otelExporter == "grpc" || otelExporter == "http" {
+	if engineEnableOtelType == "grpc" || engineEnableOtelType == "http" {
 		app.Use(otelfiber.Middleware(otelfiber.WithNext(func(c *fiber.Ctx) bool {
 			// these endpoints will be traced
-			if c.Path() == v1Path || c.Path() == roPath || (execPath != "" && c.Path() == execPath) {
+			if c.Path() == rwPath || c.Path() == roPath || (scriptingPublicPath != "" && c.Path() == scriptingPublicPath) {
 				return false
 			}
 			// skip for others endpoint
@@ -189,65 +164,43 @@ func setupFiber(startupCtx context.Context, startupReadonlyCtx context.Context, 
 		})))
 	}
 
-	// get the HEALTH_CHECK_PATH from environment variable
-	var healthCheckPath = os.Getenv("ENGINE_PLUS_HEALTH_CHECK_PATH")
-	// default to /public/graphql/health if the env is not set
-	if healthCheckPath == "" {
-		healthCheckPath = "/public/graphql/health"
-	}
 	app.Get(healthCheckPath, func(c *fiber.Ctx) error {
 		// check for startupCtx to be done
 		if startupCtx.Err() == nil {
-			log.Warn("Startup has not yet completed, return OK for health check anyway")
-			return c.SendStatus(fiber.StatusOK)
-			// Or we can wait for startupCtx to be done here: <-startupCtx.Done()
+			log.Warn("Startup has not yet completed, return health check not OK")
+			return c.SendStatus(fiber.StatusServiceUnavailable)
 		}
-		var agent *fiber.Agent
+		// startup has completed, we can now do the health check for dependent services
+		var url string
 		if startupReadonlyCtx.Err() == nil {
 			// fire GET request to scripting-server to do healthcheck of
 			// only primary-engine, as replica-engine is not yet ready
-			agent = fiber.Get("http://localhost:8880/health/engine?not=replica")
+			url = "http://localhost:8880/health/engine?quite=true&not=replica"
 		} else {
 			// fire GET request to scripting-server to do full healthcheck of all engines
-			agent = fiber.Get("http://localhost:8880/health/engine")
+			url = "http://localhost:8880/health/engine?quite=true"
 		}
-		if err := agent.Parse(); err != nil {
-			log.Error(err)
+		if err := proxy.Do(c, url); err != nil {
+			return err
 		}
-		code, body, errs := agent.Bytes()
-		if len(errs) > 0 {
-			log.Error(errs)
-			return c.Status(500).SendString(errs[0].Error())
-		}
-		// return the response from the upstream url
-		return c.Status(code).Send(body)
+		// Remove Server header from response
+		c.Response().Header.Del(fiber.HeaderServer)
+		return nil
 	})
 	// standard health check endpoint
 	app.Get("/health", func(c *fiber.Ctx) error {
-		// read GET parameter from sleep from url eg: localhost:3000/health?sleep=30
-		sleep := c.Query("sleep")
-		// ParseInt string sleep to int
-		sleepDuration, err := strconv.ParseInt(sleep, 10, 64)
-		// sleep for the given time, sleepDuration is in microseconds
-		if err == nil && sleepDuration > 0 {
-			// sleep max 110 seconds, as App Runner will always timeout after 120 seconds
-			time.Sleep(time.Duration(min(sleepDuration, 110_000_000)) * time.Microsecond)
+		// check if startupCtx is done
+		if startupCtx.Err() == nil {
+			log.Warn("Startup has not yet completed, return health check not OK")
+			return c.SendStatus(fiber.StatusServiceUnavailable)
 		}
 		return c.SendStatus(fiber.StatusOK)
 	})
 
-	primaryVsReplicaWeight := 100
-	if os.Getenv("HASURA_GRAPHQL_READ_REPLICA_URLS") != "" {
-		primaryVsReplicaWeight = 50
-		// parse the primary weight from env, convert it to int
-		primaryWeightInt, err := strconv.Atoi(os.Getenv("ENGINE_PLUS_GRAPHQL_PRIMARY_VS_REPLICA_WEIGHT"))
-		if err == nil {
-			primaryVsReplicaWeight = primaryWeightInt
-		}
-	}
-
+	mutationHandler := graphqlMutationHandlerFactory()
+	queryHandler := graphqlQueryHandlerFactory(startupReadonlyCtx, redisCacheClient)
 	// add a POST endpoint to forward request to an upstream url
-	app.Post(v1Path, func(c *fiber.Ctx) error {
+	app.Post(rwPath, func(c *fiber.Ctx) error {
 		// check and wait for startupCtx to be done
 		waitForStartupToBeCompleted(startupCtx)
 		// check if this is read only request
@@ -260,87 +213,14 @@ func setupFiber(startupCtx context.Context, startupReadonlyCtx context.Context, 
 			// TODO: support subscription via another websocket endpoint in the future
 			return fiber.NewError(fiber.StatusForbidden, "GraphQL Engine Plus does not support subscription yet")
 		}
-		// check if this is a query or mutation (can't be mutation)
-		queryType := "query"
+		// check if this is a query or mutation
 		if graphqlReq.IsMutationGraphQLRequest() {
-			queryType = "mutation"
+			// pass to read write handler since this is a graphql muration
+			return mutationHandler(c, graphqlReq)
+		} else {
+			// pass to read only handler if this is a graphql query
+			return queryHandler(c, graphqlReq)
 		}
-		var redisKey string
-		var ttl int
-		var familyCacheKey, cacheKey uint64
-		if queryType == "query" {
-			// check if this is a cached query
-			if ttl = graphqlReq.IsCachedQueryGraphQLRequest(); ttl != 0 && redisCacheClient != nil {
-				familyCacheKey, cacheKey = CalculateCacheKey(c, graphqlReq)
-				// check if the response body of this query has already cached in redis
-				// if yes, return the response from redis
-				redisKey = CreateRedisKey(familyCacheKey, cacheKey)
-				if cacheData, err := redisCacheClient.Get(c.Context(), redisKey,
-					TraceOptions{tracer, c.UserContext()}); err == nil {
-					log.Debug("Cache hit for cacheKey: ", cacheKey)
-					return SendCachedResponseBody(c, cacheData, ttl, familyCacheKey, cacheKey, TraceOptions{tracer, c.UserContext()})
-				}
-				log.Debug("Cache miss for cacheKey: ", cacheKey)
-			}
-		}
-		req := c.Request()
-		resp := c.Response()
-		// prepare the proxy request
-		prepareProxyRequest(req, "localhost:8881", "/v1/graphql", c.IP())
-		// start tracer span
-		spanCtx, span := tracer.Start(c.UserContext(), "primary-engine",
-			oteltrace.WithSpanKind(oteltrace.SpanKindInternal),
-			oteltrace.WithAttributes(
-				attribute.String("graphql.operation.name", graphqlReq.OperationName),
-				attribute.String("graphql.operation.type", queryType),
-				attribute.String("graphql.document", graphqlReq.Query),
-				attribute.String("http.request.header.x_request_id", c.Get("X-Request-ID")),
-			))
-		carrier := FastHttpHeaderCarrier{&req.Header}
-		propagators.Inject(spanCtx, carrier)
-		if err = primaryEngineServerProxyClient.DoTimeout(req, resp, UPSTREAM_TIME_OUT); err != nil {
-			log.Error("Failed to request primary-engine:", err)
-		}
-		span.End() // end tracer span
-		// process the proxy response
-		processProxyResponse(c, resp, redisCacheClient, ttl, familyCacheKey, cacheKey)
-		return err
-	})
-
-	// get the PATH from environment variable
-	var engineMetaPath = os.Getenv("ENGINE_PLUS_META_PATH")
-	// default to /public/meta/ if the env is not set
-	if engineMetaPath == "" {
-		engineMetaPath = "/public/meta/"
-	}
-	// check if metadataPath end with /
-	// if not, add / to the end of metadataPath.
-	if engineMetaPath[len(engineMetaPath)-1:] != "/" {
-		engineMetaPath = engineMetaPath + "/"
-	}
-	// add a POST endpoint to forward request to an upstream url
-	app.Post(engineMetaPath+"+", func(c *fiber.Ctx) error {
-		// check and wait for startupCtx to be done
-		waitForStartupToBeCompleted(startupCtx)
-		// fire a POST request to the upstream url using the same header and body from the original request
-		// any path after /public/meta/ will be appended to the upstream url
-		// allow any type of request to be sent to the hasura graphql engine
-		// for ex: /public/meta/v1/metadata -> /v1/metadata
-		// for ex: /public/meta/v2/query -> /v2/query
-		// read more here: https://hasura.io/docs/latest/api-reference/overview/
-		agent := fiber.Post("http://localhost:8881/" + c.Params("+"))
-		// send request to upstream without caching
-		return sendRequestToUpstream(c, agent)
-	})
-	// add a GET endpoint to forward request to an upstream url
-	app.Get(engineMetaPath+"+", func(c *fiber.Ctx) error {
-		// check and wait for startupCtx to be done
-		waitForStartupToBeCompleted(startupCtx)
-		// for ex: GET /public/meta/v1/version -> /v1/version
-		// fire a GET request to the upstream url using the same header and body from the original request
-		agent := fiber.Get("http://localhost:8881/" + c.Params("+"))
-		// send request to upstream without caching
-		return sendRequestToUpstream(c, agent)
 	})
 
 	// add a POST endpoint to forward request to an upstream url
@@ -361,89 +241,57 @@ func setupFiber(startupCtx context.Context, startupReadonlyCtx context.Context, 
 			// TODO: support subscription via another websocket endpoint in the future
 			return fiber.NewError(fiber.StatusForbidden, "GraphQL Engine Plus does not support subscription yet")
 		}
-		var redisKey string
-		var ttl int
-		var familyCacheKey, cacheKey uint64
-		if ttl = graphqlReq.IsCachedQueryGraphQLRequest(); ttl != 0 && redisCacheClient != nil {
-			familyCacheKey, cacheKey = CalculateCacheKey(c, graphqlReq)
-			// check if the response body of this query has already cached in redis
-			// if yes, return the response from redis
-			redisKey = CreateRedisKey(familyCacheKey, cacheKey)
-			if cacheData, err := redisCacheClient.Get(c.Context(), redisKey,
-				TraceOptions{tracer, c.UserContext()}); err == nil {
-				log.Debug("Cache hit for cacheKey: ", cacheKey)
-				return SendCachedResponseBody(c, cacheData, ttl, familyCacheKey, cacheKey, TraceOptions{tracer, c.UserContext()})
-			}
-			log.Debug("Cache miss for cacheKey: ", cacheKey)
+		return queryHandler(c, graphqlReq)
+	})
+
+	// Proxy request to an upstream url
+	app.All(engineMetaPath+"+", func(c *fiber.Ctx) error {
+		// check and wait for startupCtx to be done
+		waitForStartupToBeCompleted(startupCtx)
+		// Ex if engineMetaPath = "/public/meta"
+		// for ex: POST /public/meta/v1/metadata -> /v1/metadata
+		// for ex: POST /public/meta/v2/query -> /v2/query
+		// for ex: GET /public/meta/v1/version -> /v1/version
+		// proxy GET request to the upstream using the same header and body from the original request
+		url := "http://localhost:8881" + c.Params("+")
+		if err := proxy.Do(c, url); err != nil {
+			return err
 		}
-		req := c.Request()
-		resp := c.Response()
-		// random an int from 0 to 99
-		// if randomInt is less than primaryWeight, send request to primary upstream
-		// if replica is not available, always send request to primary upstream
-		if startupReadonlyCtx.Err() == nil || primaryVsReplicaWeight >= 100 || (primaryVsReplicaWeight > 0 && rand.Intn(100) < primaryVsReplicaWeight) {
-			// prepare the proxy request to primary upstream
-			prepareProxyRequest(req, "localhost:8881", "/v1/graphql", c.IP())
-			// start tracer span
-			spanCtx, span := tracer.Start(c.UserContext(), "primary-engine",
-				oteltrace.WithSpanKind(oteltrace.SpanKindInternal),
-				oteltrace.WithAttributes(
-					attribute.String("graphql.operation.name", graphqlReq.OperationName),
-					attribute.String("graphql.operation.type", "query"),
-					attribute.String("graphql.document", graphqlReq.Query),
-					attribute.String("http.request.header.x_request_id", c.Get("X-Request-ID")),
-				))
-			carrier := FastHttpHeaderCarrier{&req.Header}
-			propagators.Inject(spanCtx, carrier)
-			if err = primaryEngineServerProxyClient.DoTimeout(req, resp, UPSTREAM_TIME_OUT); err != nil {
-				log.Error("Failed to request primary-engine:", err)
-			}
-			span.End() // end tracer span
-		} else {
-			// prepare the proxy request to replica upstream
-			prepareProxyRequest(req, "localhost:8882", "/v1/graphql", c.IP())
-			// start tracer span
-			spanCtx, span := tracer.Start(c.UserContext(), "replica-engine",
-				oteltrace.WithSpanKind(oteltrace.SpanKindInternal),
-				oteltrace.WithAttributes(
-					attribute.String("graphql.operation.name", graphqlReq.OperationName),
-					attribute.String("graphql.operation.type", "query"),
-					attribute.String("graphql.document", graphqlReq.Query),
-					attribute.String("http.request.header.x_request_id", c.Get("X-Request-ID")),
-				))
-			carrier := FastHttpHeaderCarrier{&req.Header}
-			propagators.Inject(spanCtx, carrier)
-			if err = replicaEngineServerProxyClient.DoTimeout(req, resp, UPSTREAM_TIME_OUT); err != nil {
-				log.Error("Failed to request replica-engine:", err)
-			}
-			span.End() // end tracer span
-		}
-		// process the proxy response
-		processProxyResponse(c, resp, redisCacheClient, ttl, familyCacheKey, cacheKey)
-		return err
+		// Remove Server header from response
+		c.Response().Header.Del(fiber.HeaderServer)
+		return nil
 	})
 
 	// this endpoint is optional and will only be available if the env is set
-	if execPath != "" {
+	if scriptingPublicPath != "" {
 		// this endpoint expose the scripting-server execute endpoint to public
 		// this allow client to by pass GraphQL engine and execute script directly
 		// be careful when exposing this endpoint to public without a strong security measure
-		app.Post(execPath, func(c *fiber.Ctx) error {
+		app.All(scriptingPublicPath+"+", func(c *fiber.Ctx) error {
 			// validate the engine plus execute secret
-			headerHasuraAdminSecret := c.Get("X-Engine-Plus-Execute-Secret")
-			envHasuraAdminSecret := os.Getenv("ENGINE_PLUS_EXECUTE_SECRET")
+			headerExecuteSecret := c.Get("X-Engine-Plus-Execute-Secret")
 			// if header is empty or not equal to the env, return 401
-			if headerHasuraAdminSecret == "" || headerHasuraAdminSecret != envHasuraAdminSecret {
+			if headerExecuteSecret == "" || headerExecuteSecret != envExecuteSecret {
 				return fiber.NewError(fiber.StatusUnauthorized, "Unauthorized")
+			}
+			// check if the header X-Engine-Plus-Execute-Url is present
+			// when allowExecuteUrl is false, if yes return 403
+			if !allowExecuteUrl && c.Get("X-Engine-Plus-Execute-Url") != "" {
+				return fiber.NewError(fiber.StatusForbidden, "Forbidden")
 			}
 			// check and wait for startupCtx to be done
 			waitForStartupToBeCompleted(startupCtx)
+			// Ex if engineMetaPath = "/scripting"
+			// for ex: POST /scripting/upload -> /upload
+			// for ex: POST /scripting/execute -> /execute
+			// for ex: GET /scripting/validate -> /validate
+			// for ex: GET /scripting/health/engine -> /health/engine
 			req := c.Request()
 			resp := c.Response()
 			// prepare the proxy request
-			prepareProxyRequest(req, "localhost:8880", "/execute", c.IP())
+			prepareProxyRequest(req, "localhost:8880", c.Params("+"), c.IP())
 			// start tracer span
-			spanCtx, span := tracer.Start(c.UserContext(), "proxy -> scripting-server: /execute",
+			spanCtx, span := tracer.Start(c.UserContext(), "proxy -> scripting-server",
 				oteltrace.WithSpanKind(oteltrace.SpanKindInternal),
 				oteltrace.WithAttributes(
 					attribute.String("http.request.header.x_request_id", c.Get("X-Request-ID")),
@@ -465,56 +313,58 @@ func setupFiber(startupCtx context.Context, startupReadonlyCtx context.Context, 
 			return err
 		})
 	}
+
 	return app
 }
 
-var redisUrlRemoveUnexpectedOptionRegex = regexp.MustCompile(`(?i)(\?)?&?(ssl_cert_reqs=\w+)`)
-
 func setupRedisClient(ctx context.Context) *RedisCacheClient {
 	// setup redis client
-	redisUrl := os.Getenv("HASURA_GRAPHQL_REDIS_CLUSTER_URL")
-	if redisUrl == "" {
-		redisUrl = os.Getenv("HASURA_GRAPHQL_REDIS_URL")
+
+	if hasuraGqlRedisUrl == "" && hasuraGqlRedisClusterUrl == "" {
+		return nil
 	}
 	groupcacheOptions := NewGroupCacheOptions()
-	// parse groupcacheOptions.waitETA from env
-	if waitETA, err := strconv.ParseInt(os.Getenv("ENGINE_PLUS_GROUPCACHE_WAIT_ETA"), 10, 64); err == nil && waitETA >= 0 {
-		groupcacheOptions.waitETA = uint64(waitETA)
-	}
+	// set groupcacheOptions.waitETA using value from env
+	groupcacheOptions.waitETA = engineGroupcacheWaitEta
+	groupcacheOptions.cacheMaxSize = int64(engineGroupcacheMaxSize)
 
 	// get list of available addresses
 	localAddress := "127.0.0.1"
-	if os.Getenv("ENGINE_PLUS_GROUPCACHE_CLUSTER_MODE") != "localhost" {
-		addrs, err := net.InterfaceAddrs()
+	if engineGroupcacheClusterMode == "aws_ec2" {
+		// get the local ipv4 address of the ec2 instance via the meta-data url
+		address, err := GetIpFromAwsEc2Metadata()
 		if err != nil {
 			log.Error(err)
 		} else {
-			for _, addr := range addrs {
-				if ipnet, ok := addr.(*net.IPNet); ok && !ipnet.IP.IsLoopback() {
-					// check if IPv4 or IPv6 is not nil
-					if ipnet.IP.To4() != nil {
-						// get the first ipv4 address to set as localserver address
-						localAddress = ipnet.IP.String()
-						break
-					}
-					if ipnet.IP.To16() != nil {
-						// may skip ipv6 for now
-						log.Debug(ipnet.IP.String())
-					}
-				}
-			}
+			localAddress = address
+		}
+	} else if engineGroupcacheClusterMode == "aws_ecs" {
+		// get the local ipv4 address of the container instance via the ecs meta-data
+		address, err := GetIpFromAwsEcsContainerMetadata()
+		if err != nil {
+			log.Error(err)
+		} else {
+			localAddress = address
+		}
+	} else if engineGroupcacheClusterMode == "host_ipnet" {
+		address, err := GetIpFromHostNetInterfaces()
+		if err != nil {
+			log.Error(err)
+		} else {
+			localAddress = address
 		}
 	} else {
 		groupcacheOptions.disableAutoDiscovery = true
 	}
-	groupcacheServerPort := os.Getenv("ENGINE_PLUS_GROUPCACHE_PORT")
+
 	if groupcacheServerPort == "" {
 		groupcacheServerPort = "8879"
 	}
 	redisCacheClient, err := NewRedisCacheClient(
 		ctx,
-		redisUrl,
-		os.Getenv("HASURA_GRAPHQL_REDIS_READER_URL"),
+		hasuraGqlRedisClusterUrl,
+		hasuraGqlRedisUrl,
+		hasuraGqlRedisReaderUrl,
 		"http://"+localAddress+":"+groupcacheServerPort,
 		groupcacheOptions,
 	)
@@ -525,20 +375,25 @@ func setupRedisClient(ctx context.Context) *RedisCacheClient {
 	// ping redis server
 	if err := redisCacheClient.redisClient.Ping(ctx).Err(); err != nil {
 		log.Error("Failed to connect to Redis Server: ", err)
+		return nil
 	}
 	if err := redisCacheClient.redisClientReader.Ping(ctx).Err(); err != nil {
 		log.Error("Failed to connect to Redis Reader Server: ", err)
+		return nil
 	}
 	log.Info("Connected to Redis Cache")
 	return redisCacheClient
 }
 
 func main() {
+	// Init config
+	InitConfig()
+	// These context are for startup only
 	mainCtx, mainCtxCancelFn := context.WithCancel(context.Background())
 	startupCtx, startupDoneFn := context.WithTimeout(mainCtx, 60*time.Second)
 	startupReadonlyCtx, startupReadonlyDoneFn := context.WithTimeout(mainCtx, 180*time.Second)
 	// this must happen as early as possible
-	otelTracerProvider := InitTracerProvider(mainCtx, otelExporter)
+	otelTracerProvider := InitTracerProvider(mainCtx, engineEnableOtelType)
 	// setup resources
 	redisCacheClient := setupRedisClient(mainCtx)
 	// these 2 line need to happen after InitTracerProvider
@@ -546,16 +401,22 @@ func main() {
 	propagators = otel.GetTextMapPropagator()
 	// setup app
 	app := setupFiber(startupCtx, startupReadonlyCtx, redisCacheClient)
-	// get the server Host:Port from environment variable
-	var serverHost = os.Getenv("ENGINE_PLUS_SERVER_HOST")
-	// default to empty string if the env is not set
-	var serverPort = os.Getenv("ENGINE_PLUS_SERVER_PORT")
-	// default to 8000 if the env is not set
-	if serverPort == "" {
-		serverPort = "8000"
+
+	if engineServerPort == "" {
+		engineServerPort = "8000"
 	}
-	// start the http server
-	go app.Listen(serverHost + ":" + serverPort)
+
+	// start the http server in coroutine and handler error
+	// if the server failed to start
+	// we dont need to wait for the server to start
+	// as we will wait for the startup to be completed
+	go func() {
+		err := app.Listen(engineServerHost + ":" + engineServerPort)
+		if err != nil {
+			// trigger graceful shutdown
+			mainCtxCancelFn()
+		}
+	}()
 	// start the engine servers
 	serverShutdownErrorChanel := make(chan error)
 	startServerCtx, startServerCtxCancelFn := context.WithCancel(mainCtx)
@@ -563,14 +424,13 @@ func main() {
 
 	// register engine-servers & http-server clean-up operations
 	// then wait for termination signal to do the clean-up
-	shutdownTimeout := 30 * time.Second
 	cleanUpOps := map[string]gfshutdown.Operation{
 		"engine-servers": func(shutdownCtx context.Context) error {
 			startServerCtxCancelFn()
 			return <-serverShutdownErrorChanel
 		},
 		"http-server": func(shutdownCtx context.Context) error {
-			return app.ShutdownWithTimeout(shutdownTimeout - 1*time.Second)
+			return app.ShutdownWithTimeout(GRACEFUL_SHUTDOWN_TIMEOUT - 1*time.Second)
 		},
 		// Add other cleanup operations here
 	}
@@ -585,13 +445,13 @@ func main() {
 			}
 		}
 	}
-	if otelExporter == "grpc" || otelExporter == "http" {
+	if engineEnableOtelType == "grpc" || engineEnableOtelType == "http" {
 		// add opentelemetry tracer clean-up operation
 		cleanUpOps["opentelemetry-tracer"] = func(shutdownCtx context.Context) error {
 			return otelTracerProvider.Shutdown(shutdownCtx)
 		}
 	}
-	wait := gfshutdown.GracefulShutdown(mainCtx, shutdownTimeout, cleanUpOps)
+	wait := gfshutdown.GracefulShutdown(mainCtx, GRACEFUL_SHUTDOWN_TIMEOUT, cleanUpOps)
 	exitCode := <-wait
 	log.Info("GraphQL Engine Plus is shutdown")
 	os.Exit(exitCode)
