@@ -3,16 +3,16 @@ package main
 import (
 	"context"
 	"log"
+	"time"
 
 	"github.com/valyala/fasthttp"
 	"go.opentelemetry.io/otel/sdk/resource"
 
-	xray "go.opentelemetry.io/contrib/propagators/aws/xray"
+	"go.opentelemetry.io/contrib/propagators/autoprop"
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/exporters/otlp/otlptrace"
 	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc"
 	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracehttp"
-	"go.opentelemetry.io/otel/propagation"
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 	semconv "go.opentelemetry.io/otel/semconv/v1.4.0"
 	oteltrace "go.opentelemetry.io/otel/trace"
@@ -23,26 +23,28 @@ type TraceOptions struct {
 	ctx    context.Context
 }
 
-func InitTracerProvider(ctx context.Context, otelExporter string) *sdktrace.TracerProvider {
+func InitTracerProvider(ctx context.Context, otelTracerType string) *sdktrace.TracerProvider {
 	var exporter *otlptrace.Exporter
 	var err error
-	if otelExporter == "http" {
+	if otelTracerType == "http" {
 		client := otlptracehttp.NewClient()
 		exporter, err = otlptrace.New(ctx, client)
-	} else if otelExporter == "grpc" {
+	} else if otelTracerType == "grpc" {
 		exporter, err = otlptracegrpc.New(ctx)
 	} else {
-		if otelExporter != "" && otelExporter != "false" {
-			log.Println("Error, unknown Open Telemetry exporter: ", otelExporter)
+		if otelTracerType != "" && otelTracerType != "false" {
+			log.Println("Error, unknown Open Telemetry exporter: ", otelTracerType)
 		}
 		tp := sdktrace.NewTracerProvider()
-		tp.Shutdown(ctx)
+		// calling Shutdown create a noops tracer provider
+		_ = tp.Shutdown(ctx)
 		return tp
 	}
 	if err != nil {
 		log.Println("Error when init Open Telemetry tracer: ", err)
 		tp := sdktrace.NewTracerProvider()
-		tp.Shutdown(ctx)
+		// calling Shutdown create a noops tracer provider
+		_ = tp.Shutdown(ctx)
 		return tp
 	}
 	tp := sdktrace.NewTracerProvider(
@@ -54,8 +56,28 @@ func InitTracerProvider(ctx context.Context, otelExporter string) *sdktrace.Trac
 			)),
 	)
 	otel.SetTracerProvider(tp)
-	otel.SetTextMapPropagator(propagation.NewCompositeTextMapPropagator(propagation.TraceContext{}, xray.Propagator{}))
+	otel.SetTextMapPropagator(autoprop.NewTextMapPropagator())
 	return tp
+}
+
+func WrapContextCancelByAnotherContext(mainCtx context.Context, triggerCtx context.Context, timeoutDuration time.Duration) (newCtx context.Context, cancel context.CancelFunc) {
+	// create a new context with cancel
+	if timeoutDuration > 0 {
+		newCtx, cancel = context.WithTimeout(mainCtx, timeoutDuration*time.Millisecond)
+	} else {
+		newCtx, cancel = context.WithCancel(mainCtx)
+	}
+	// create a new goroutine to cancel the context
+	go func() {
+		select {
+		case <-triggerCtx.Done():
+			cancel()
+		case <-newCtx.Done():
+			// pass
+		}
+	}()
+	// return ctx, cancel
+	return
 }
 
 // FastHttpHeaderCarrier adapts fasthttp.RequestHeader to satisfy the TextMapCarrier interface.
